use splash screen to hide initialization
