+++ flutter_native_splash.yaml
@@ -1,0 +1,14 @@
+flutter_native_splash:
+  color: "#42a5f5"
+  image: assets/carewool-icon.png
+  color_dark: "#042a49"
+  image_dark: assets/carewool-icon.png
+
+  android_12:
+    image: assets/carewool-icon.png
+    icon_background_color: "#ffffff"
+    color: "#42a5f5"
+
+  ios: false
+  web: false
+  fullscreen: true
