+++ lib/navigation/app_router.dart
@@ -2,7 +2,7 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
 import '../widget/page/calculator_page.dart';
-import '../widget/page/loading_screen.dart';
+import '../widget/page/default_page.dart';
 
 part 'app_router.gr.dart';
 
@@ -10,8 +10,7 @@ part 'app_router.gr.dart';
 class AppRouter extends _$AppRouter {
   @override
   List<AutoRoute> get routes => [
-        AutoRoute(page: LoadingRoute.page, initial: true),
-        RedirectRoute(path: '/', redirectTo: '/calc'),
+        AutoRoute(page: DefaultRoute.page, initial: true),
         AutoRoute(path: '/calc', page: CalculatorRoute.page),
       ];
 }
