+++ pubspec.yaml
@@ -28,6 +28,7 @@ dependencies:
   path: ^1.8.3
   intl: ^0.18.0
   auto_route: ^6.3.0
+  flutter_native_splash: ^2.2.19
 
 dev_dependencies:
   flutter_test:
@@ -40,3 +41,5 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
+  assets:
+    - assets/
